airflow/requirements.txt
requests
#pyspark
google-cloud-storage
apache-airflow-providers-google
pyarrow
pandas
duckdb
dbt-bigquery